feat: populated controllers, changes to routing_diag, rm user module
The controllers are populated with the necessary routes (still temporary), we'll further proceed on populating services in a test-driven development environment. Firstly, we'll be completing the Auth, token signing and Passport business logic. I have also removed the user module which is totally unnecessary. The routing diagram has been updated with removal of an optional route and further a typo fix.