feat: api.ts, zustand, react-router and protection
Added api.ts as AxiosConfig routing to the backend URL. ENVs are now activated. Zustand is integrated with Zustand Store having its own useAuthStore, which tracks the user, token, checksAuth, login, logout and checks if the user is logged in, which is defined by setting the user & jwtToken by the checkAuth() which is repeatedly called under ProtectedRoute.ts. Protected Route for now only wraps /dashboard, we'll enable Auth then wrap the remaining routes