feat: interceptor - response and cache
Interceptors active. Response is now sent in a well structured manner with statusCode and timestamp. Caching is finally introducted with advanced modifications, where we don't store to cache unauth user data as well as we store only GET requests. Furthermore advancement to caching would be done when Redis is added to the stack, next hit is up for Exception Filters then Custom Decorators, after it we finally move to frontend